Store found entrances on the server
